Add confirmation prompt

+++ lua/telescope/actions/init.lua
@@ -295,6 +295,10 @@ end
 actions.git_delete_branch = function(prompt_bufnr)
   local cwd = action_state.get_current_picker(prompt_bufnr).cwd
   local selection = action_state.get_selected_entry()
+
+  local confirmation = vim.fn.input('Do you really wanna delete branch ' .. selection.value .. '? [Y/n] ')
+  if confirmation ~= '' and string.lower(confirmation) ~= 'y' then return end
+
   actions.close(prompt_bufnr)
   local _, ret, stderr = utils.get_os_command_output({ 'git', 'branch', '-D', selection.value }, cwd)
   if ret == 0 then
@@ -311,6 +315,10 @@ end
 actions.git_rebase_branch = function(prompt_bufnr)
   local cwd = action_state.get_current_picker(prompt_bufnr).cwd
   local selection = action_state.get_selected_entry()
+
+  local confirmation = vim.fn.input('Do you really wanna delete branch ' .. selection.value .. '? [Y/n] ')
+  if confirmation ~= '' and string.lower(confirmation) ~= 'y' then return end
+
   actions.close(prompt_bufnr)
   local _, ret, stderr = utils.get_os_command_output({ 'git', 'rebase', selection.value }, cwd)
   if ret == 0 then
